MrImage.plot: changed default such that 'useSlider' has now a 4D image as default
git-svn-id: http://chap04.ethz.ch/fmri_svn/development@486 af1c8887-3614-4b1b-9890-8700dbf8cd7f

+++ analysis/code/classes/@MrImage/plot.m
@@ -234,15 +234,30 @@ if any(plotDataSpecified)
     selectStr = varargin(plotDataSpecified);
     [plotImage, ~, ~] = plotImage.select('type', selectionType, ...
         selectStr{:});
-else % 1 image with all samples of first three dimensions, for all further
-    % dimensions only first image is plotted
-    if plotImage.dimInfo.nDims > 3
-        nDimsSelect = plotImage.dimInfo.nDims - 3;
-        dimLabelsSelect = plotImage.dimInfo.dimLabels;
-        selectStr(1:2:nDimsSelect*2) = dimLabelsSelect(4:end);
-        selectStr(2:2:nDimsSelect*2) = {1};
-        plotImage = plotImage.select('type', selectionType, ...
-            selectStr{:});
+else
+    if ~useSlider % default: no slider used
+        % 1 image with all samples of first three dimensions, for all further
+        % dimensions only first sample is plotted
+        if plotImage.dimInfo.nDims > 3
+            nDimsSelect = plotImage.dimInfo.nDims - 3;
+            dimLabelsSelect = plotImage.dimInfo.dimLabels;
+            selectStr(1:2:nDimsSelect*2) = dimLabelsSelect(4:end);
+            selectStr(2:2:nDimsSelect*2) = {1};
+            plotImage = plotImage.select('type', selectionType, ...
+                selectStr{:});
+        end
+    else % use slider
+        % 1 image with all samples of first FOUR dimensions, for all further
+        % dimensions only first sample is plotted
+        if plotImage.dimInfo.nDims > 4
+            nDimsSelect = plotImage.dimInfo.nDims - 4;
+            dimLabelsSelect = plotImage.dimInfo.dimLabels;
+            selectStr(1:2:nDimsSelect*2) = dimLabelsSelect(5:end);
+            selectStr(2:2:nDimsSelect*2) = {1};
+            plotImage = plotImage.select('type', selectionType, ...
+                selectStr{:});
+        end
+        
     end
 end
 
@@ -666,14 +681,15 @@ else % different plot types: montage, 3D, spm
                     delete(fileNameDimInfo);
                     [stat, mess, id] = rmdir(this.parameters.save.path);
                 end
-                
-                for iAddImages = 1:nAddImages
-                    if doDeleteAddImages{iAddImages}
-                        delete(fileNameNiftiAddImages{iAddImages});
-                        [fp, fn] = fileparts(fileNameNiftiAddImages{iAddImages});
-                        fileNameDimInfo = fullfile(fp, [fn '_dimInfo.mat']);
-                        delete(fileNameDimInfo);
-                        [stat, mess, id] = rmdir(overlayImages{iAddImages}.parameters.save.path);
+                if doPlotAdditionalImages
+                    for iAddImages = 1:nAddImages
+                        if doDeleteAddImages{iAddImages}
+                            delete(fileNameNiftiAddImages{iAddImages});
+                            [fp, fn] = fileparts(fileNameNiftiAddImages{iAddImages});
+                            fileNameDimInfo = fullfile(fp, [fn '_dimInfo.mat']);
+                            delete(fileNameDimInfo);
+                            [stat, mess, id] = rmdir(overlayImages{iAddImages}.parameters.save.path);
+                        end
                     end
                 end
             end
